Fixed orientation of graph
The overall graph was upside down, so the node order on edges was
flipped around, making Graphviz draw the graph the other way up

This meant that the arrows pointed in the wrong direction, so I flipped
that.

example_data.csv had a problem in transferring from Windows to Linux,
and quotation marks were changed to question marks. This has been fixed.