Refactored open_csv.py to use an object-oriented approach
open_csv.py used to simply create a list of lists. Each sublist was a
node and contained the node's attributes in a specific order.

I could have replaced this with a dict, but chose to create a Node class
for more clarity. This means I can address a node's attributes by name,
rather than by an unmemorable index.